Add files via upload
In this Excel file I show a model for the budget of a firm, taking into account various factors. In the first workbook are modeled the income statement, the balance sheet and cash flows along with the calculations of NPV, IRR, payback period and other indicators that represent the health of the firm. In the second workbook of the file, I model the capital expenditures of the firm, used to complete the calculations of the first workbook. In the third I calculate the WACC, in the fourth I calculate the Economic Value Added of the firm.